fixed MyAcc methods

+++ app/src/main/java/com/example/user/loginwhithfb/model/UserLoginInfoTable.java
@@ -80,7 +80,7 @@ public class UserLoginInfoTable {
         return mobileNumber;
     }
 
-    public void setMobileNumber(int mobileNumber) {
+    public void setMobileNumber(long mobileNumber) {
         this.mobileNumber = mobileNumber;
     }
